test(e2e): tier-gate Premium coverage instead of runtime-skipping on CE
TestMeta_GroupHookExtras (group webhooks) and the MR dependencies
coverage are Premium features: on CE they could only ever skip, first
via a 404/unknown-action sniff and, since the catalog gained per-tier
action enums, via schema-validation text matching. Move them to the
enterprise build like the rest of the EE coverage:

- TestMeta_GroupHookExtras and its helpers now live in
  groupextras_ee_test.go with the tier-skip removed — on EE the
  hook_add fixture must succeed.
- The DependenciesList subtest of TestIndividual_MRExtras becomes
  TestIndividual_MRDependenciesList in mrextras_ee_test.go with strict
  assertions.
- network_endpoints_ce_test.go becomes network_endpoints_test.go
  (build tag e2e): the fixture-service URL helpers are edition-agnostic
  and the EE webhook coverage needs them too.
